+++ larcv/app/ThreadIO/BatchFillerTensor3D.cxx
@@ -31,7 +31,7 @@ void BatchFillerTensor3D::configure(const PSet& cfg) {
 void BatchFillerTensor3D::initialize() {}
 
 void BatchFillerTensor3D::_batch_begin_() {
-  if(!batch_data().dim().empty() && batch_size() != batch_data().dim().front()) {
+  if(!batch_data().dim().empty() && (int)(batch_size()) != batch_data().dim().front()) {
     LARCV_INFO() << "Batch size changed " << batch_data().dim().front() << "=>" << batch_size() << std::endl;
     auto dim = batch_data().dim();
     dim[0] = batch_size();
